requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
gunicorn==21.2.0
numpy==1.26.2
scipy==1.11.4
